docker/requirements.txt
selenium==3.141.0
pyvirtualdisplay
beautifulsoup4==4.10.0
pandas==1.3.5
discord.py==2.0.1